refactor: Move AddTeamMemberForm to new file
